NRR-webpage/requirements.txt
streamlit>=1.35.0
plotly>=5.20.0
pandas>=2.0.0
numpy>=1.26.0
faiss-cpu>=1.7.4
sentence-transformers>=2.6.0
openai>=1.30.0
duckduckgo-search>=6.1.0
